+++ src/main/java/egg/finalproject/order/OrderService.java
@@ -1,8 +1,17 @@
 package egg.finalproject.order;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class OrderService {
-
+	@Autowired
+	private OrderDAO dao;
+	
+	//TempOrderDTO : ProductDTO(상품정보) + quantity(구매 수량)
+		//(product_no에 해당하는 상품 정보 + 고객 구매 희망 수량)
+		//RETURN : TempOrderDTO반환
+		public TempOrderDTO getTempOrderInfo (int product_no) throws Exception{
+			return dao.getTempOrderInfo(product_no);
+		}
 }
