+++ src/main/java/egg/finalproject/product/ProductController.java
@@ -69,19 +69,5 @@ public class ProductController {
 		return list;
 	}
 	
-	//주문 페이지로 이동
-	//TempOrderDTO : ProductDTO + quantity 
-	//RETURN : 주문페이지 주소값 반환 (TempOrderDTO 보냄) 
-	@RequestMapping("/toOrder.do")
-	public String toOrder(int product_no,int quantity, Model model) throws Exception{
-		System.out.println("product_no : "+product_no);
-		System.out.println("quantity : "+quantity);
-		TempOrderDTO dto = service.getTempOrderInfo(product_no); //product_no에 해당하는 ProductDTO정보를 먼저 담고
-		dto.setQuantity(quantity); // 수량을 같이 담아서 넘긴다.
-		System.out.println(dto);
-		List<TempOrderDTO> list = new ArrayList<>();  //장바구니와 같은 페이지로 넘기기 때문에 같이 list로 맞춰주기 위함.
-		list.add(dto);
-		model.addAttribute("list",list);
-		return "order/order.jsp";
-	}
+	
 }
