+++ src/main/java/egg/finalproject/product/ProductService.java
@@ -29,12 +29,6 @@ public class ProductService {
 		return dao.getProduct(product_no);
 	}
 	
-	//TempOrderDTO : ProductDTO(상품정보) + quantity(구매 수량)
-	//(product_no에 해당하는 상품 정보 + 고객 구매 희망 수량)
-	//RETURN : TempOrderDTO반환
-	public TempOrderDTO getTempOrderInfo (int product_no) throws Exception{
-		return dao.getTempOrderInfo(product_no);
-	}
 	
 	//상품 페이지에 입력된 상품 정보 + 썸네일 저장
 	//RETURN XX
