+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -18,6 +18,7 @@
 	<resources mapping="/css/**" location="/resources/css/" />
 	<resources mapping="/file/**" location="/resources/file/" />
 	<resources mapping="/onlinePostThumbNail/**" location="/onlinePostThumbNail/"/>
+	<resources mapping="/productThumbnail/**" location="/productThumbnail/"/>
 	<resources mapping="/profilePhotos/**" location="/profilePhotos/"/>  
 
 	<!-- Resolves views selected for rendering by @Controllers to .jsp resources in the /WEB-INF/views directory -->
