+++ src/main/webapp/WEB-INF/views/product/productDetail.jsp
@@ -258,7 +258,7 @@
         	let product_no = ${ProductDTO.product_no}
         	let quantity = $("#quantityController").val();
         	console.log(product_no);
-        	location.href="/product/toOrder.do?product_no="+product_no+"&quantity="+quantity;
+        	location.href="/order/toOrder.do?product_no="+product_no+"&quantity="+quantity;
         })
     </script>
 </body>
